add items part 1

+++ LostFoundServer-ejb/src/java/model/User.java
@@ -11,12 +11,13 @@ package model;
  */
 
 public class User {
+    
     private int id;
     private String username;
     private String password;
     private String telephone;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
